libs/vizu_mcp_commons/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vizu_mcp_commons"
version = "0.1.0"
description = "Common MCP utilities for Vizu services - auth, dependencies, middleware"
authors = ["Vizu Team"]
readme = "README.md"
packages = [{include = "vizu_mcp_commons", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.0"
pydantic-settings = "^2.6.0"
pyjwt = "^2.8.0"
redis = "^5.0.0"

# Vizu dependencies
vizu_models = {path = "../vizu_models", develop = true}
vizu_db_connector = {path = "../vizu_db_connector", develop = true}
vizu_context_service = {path = "../vizu_context_service", develop = true}
vizu_auth = {path = "../vizu_auth", develop = true}
vizu_tool_registry = {path = "../vizu_tool_registry", develop = true}

# Optional MCP integration
fastmcp = {version = "^2.11.0", optional = true}
fastapi = {version = "^0.100.0", optional = true}

[tool.poetry.extras]
mcp = ["fastmcp"]
fastapi = ["fastapi"]
full = ["fastmcp", "fastapi"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
